productor_notificaciones/requirements.txt
google-cloud-tasks==2.7.1
Click==7.0
flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0